+++ scheduling/scheduler.slascheduler/src/de/uniol/inf/is/odysseus/scheduler/slascheduler/IPriorityFunction.java
@@ -3,6 +3,6 @@ package de.uniol.inf.is.odysseus.scheduler.slascheduler;
 
 public interface IPriorityFunction {
 	
-	public int calcPriority(int... cost);
+	public double calcPriority(double... cost);
 
 }
