+++ scheduling/scheduler.slascheduler/src/de/uniol/inf/is/odysseus/scheduler/slascheduler/ISLAConformance.java
@@ -2,7 +2,7 @@ package de.uniol.inf.is.odysseus.scheduler.slascheduler;
 
 public interface ISLAConformance {
 
-	public int getConformance();
+	public double getConformance();
 	
 	public void reset();
