+++ scheduling/scheduler.slascheduler/src/de/uniol/inf/is/odysseus/scheduler/slascheduler/IStarvationFreedom.java
@@ -2,6 +2,6 @@ package de.uniol.inf.is.odysseus.scheduler.slascheduler;
 
 public interface IStarvationFreedom {
 
-	public int sf();
+	public double sf();
 	
 }
