+++ scheduling/scheduler.slascheduler/src/de/uniol/inf/is/odysseus/scheduler/slascheduler/sf/ElementTimeStampSF.java
@@ -5,7 +5,7 @@ import de.uniol.inf.is.odysseus.scheduler.slascheduler.IStarvationFreedom;
 public class ElementTimeStampSF implements IStarvationFreedom {
 
 	@Override
-	public int sf() {
+	public double sf() {
 		// TODO not implemented yet
 		return 0;
 	}
